Animations work now

+++ Renderer.js
@@ -5,24 +5,25 @@ class Renderer
 {
     transform;
     sprite;
+    spriteSheet;
+    currentAnimation;
     stage;
 
-    constructor(sprite, stage, transform = new Transform())
+    constructor(spriteSheet, stage, transform = new Transform())
     {
         if (stage == undefined)
             throw new Error('Renderer stage undefined! Was the renderer initialized correctly?');
 
-        this.sprite = sprite;
+        this.spriteSheet = spriteSheet;
         this.stage = stage;
         this.transform = transform;
+        this.setAnimation('default');
+        this.sprite.textures[0].baseTexture.scaleMode = PIXI.SCALE_MODES.NEAREST;
         this.link();
     }
         
     link()
     {
-        if (this.sprite == undefined)
-            throw new Error("Renderer sprite is undefined! Was there an error loading it?");
-        
         this.stage.addChild(this.sprite);
     }
     unlink()
@@ -31,18 +32,24 @@ class Renderer
     }
     dispose()
     {
+        if (this.sprite == undefined) return;
+
         this.unlink();
         this.sprite.destroy(true);
     }
-    updateSprite(newTexture)
+    setAnimation(animationId, speed=1, loop=false)
     {
-        this.sprite.texture = newTexture;
+        this.currentAnimation = animationId;
+        this.unlink();
+        this.sprite = new PIXI.AnimatedSprite(this.spriteSheet.animations[animationId]);
+        this.link();
+        this.sprite.loop = loop;
+        this.sprite.animationSpeed = speed;
+        this.sprite.play();
     }
     update(delta)
     {
-        var pos = this.transform.position;
-        this.sprite.x = pos.x;
-        this.sprite.y = pos.y;
-        this.sprite.rotation = this.transform.rotation;
+        var t = this.transform;
+        this.sprite.setTransform(t.position.x, t.position.y, t.scale.x, t.scale.y, t.rotation);
     }
 }
